--- src/features/index.ts
@@ -1,4 +1,0 @@
-export { LoginPage } from './login/LoginPage';
-export { ProfilePage } from './profile/ProfilePage';
-export { RegisterPage } from './register/RegisterPage'
-export { MatchsicFrame } from './home/MatchsicFrame';
