+++ src/features/utils/index.ts
@@ -1,0 +1,4 @@
+export { LoginPage } from '../login/LoginPage';
+export { ProfilePage } from '../profile/ProfilePage';
+export { RegisterPage } from '../register/RegisterPage'
+export { MatchsicFrame } from '../home/MatchsicFrame';
